Finish middle sign-up form section
Used flexbox with a flex direction of column to pile up each label with
its input, then used flexbox again to form a row of 2 fields. Finally made use
of :focus, :invalid and :valid selectors to alter the style of the input's border
whenever the user focus on a field and enters a valid or invalid input in the password
type fields.